Update README.md
NOTICE
It is important to note that during testing, errors may occur that are not related to the code or test cases. This happens due to testing on a live website that is not under our control. Additionally, it's possible that some items may not exist, have been modified, or added in the meantime.